requirements.txt
pandas
Pillow
plotly
dash
jupyter-dash
numpy
statsmodels
opencv
